Added refetch-functions to orders
