+++ components/stores/dashboard/orders/deleteOrder.tsx
@@ -1,9 +1,12 @@
 import { useParams } from 'next/navigation';
 import axios from 'axios';
 import toast from 'react-hot-toast';
+import { useOrders } from './zustand/ordersState';
 
 const DeleteOrder = (orderId: string) => {
-  const storeID = useParams();
+  const {storeID} = useParams();
+  const reFetchCompleteOrders = useOrders((state) => state.reFetchCompleteOrders);
+
   return (
     <>
       <div
@@ -13,9 +16,7 @@ const DeleteOrder = (orderId: string) => {
             .delete(`${process.env.NEXT_PUBLIC_BACKEND_URL}/api/${storeID}/orders/${orderId}`)
             .then((res) => {
               toast.success('Order deleted');
-              setTimeout(() => {
-                window.location.reload();
-              }, 1500);
+              reFetchCompleteOrders(Array.isArray(storeID) ? storeID.toString() : storeID)
               return res.data;
             })
             .catch((error) => {
